refactor: move impl snapshot4 from stream for impl_/stream
