+++ src/impl_/stream.rs
@@ -11,6 +11,7 @@ use crate::impl_::stream_loop::StreamLoop;
 use crate::impl_::stream_sink::StreamSink;
 use crate::lambda2;
 use crate::IsLambda3;
+use crate::IsLambda4;
 
 use parking_lot::Mutex;
 use parking_lot::RwLock;
@@ -281,6 +282,38 @@ impl<A: Send + 'static> Stream<A> {
         self.snapshot(
             cb,
             lambda2(move |a: &A, b: &B| f.call(a, b, &cc.sample()), deps),
+        )
+    }
+
+    pub fn snapshot4<B, C, D, E, FN>(
+        &self,
+        cb: &Cell<B>,
+        cc: &Cell<C>,
+        cd: &Cell<D>,
+        mut f: FN,
+    ) -> Stream<E>
+    where
+        B: Send + Clone + 'static,
+        C: Send + Clone + 'static,
+        D: Send + Clone + 'static,
+        E: Send + Clone + 'static,
+        FN: IsLambda4<A, B, C, D, E> + Send + Sync + 'static,
+    {
+        let mut deps = if let Some(deps2) = f.deps_op() {
+            deps2.clone()
+        } else {
+            Vec::new()
+        };
+        let cc = cc.clone();
+        let cd = cd.clone();
+        deps.push(cc.to_dep());
+        deps.push(cd.to_dep());
+        self.snapshot(
+            cb,
+            lambda2(
+                move |a: &A, b: &B| f.call(a, b, &cc.sample(), &cd.sample()),
+                deps,
+            ),
         )
     }
